Added two ROM commands, CRC8 and read-scratchpad
Added functions for:
  - Searching for raised alarms
  - Skipping ROM (addressing all slaves simultaneously)
  - Reading parasitic power supply status
  - Reading the scratchpad of a device (at least the DS18B20)
  - Calculating the one-wire 8-bit CRC with polynomial =
      = X^8 + X^5 + X^4 + 1

Created a separate function for the search algorithm since it is shared
by both the search ROM and alert search ROM commands.

Added a struct type for the one-wire devices that can contain the ROM,
scratchpad and status (as well as what bits are actually up to date).